requirements.txt
cryptography>=3.4.0
pyOpenSSL>=20.0.0
python-dotenv>=0.19.0
click>=8.0.0
pystun3>=1.0.0
tqdm>=4.62.0
